pyproject.toml
[project]
name = "financial-audit-agent"
version = "2.0.0"
description = "Agentic AI Financial Auditor - Ledger Screening, Form 16 & GST Filing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # AI / LangChain
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "openai>=1.30.0",
    # FastAPI backend
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    # Data processing
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "xlrd>=2.0.1",
    "pypdf>=4.3.0",
    "tabulate>=0.9.0",
    # PDF generation
    "reportlab>=4.2.0",
    "pillow>=10.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "backend"]

[tool.ruff]
line-length = 100
target-version = "py311"
